Refactor DHT and network code structure

+++ include/dht_hunter/network/udp_client.hpp
@@ -5,6 +5,7 @@
 #include <string>
 #include <vector>
 #include <memory>
+#include <sys/types.h>  // for ssize_t
 
 namespace dht_hunter::network {
 
@@ -48,7 +49,7 @@ public:
      * @param port The server's port.
      * @return The number of bytes sent, or -1 on error.
      */
-    int send(const std::vector<uint8_t>& data, const std::string& address, uint16_t port);
+    ssize_t send(const std::vector<uint8_t>& data, const std::string& address, uint16_t port);
 
     /**
      * @brief Set a callback function to be called when data is received.
